Adding Link Controller Minimal Setup Example
