+++ examples/configs/src/index.js
@@ -68,6 +68,7 @@ import { cellNeighborhood } from './view-configs/3d-maps/cell-neighborhood-named
 import { threeMinimal } from './view-configs/3d-maps/three-minimal.js';
 import { threeMinimalLight } from './view-configs/3d-maps/three-minimal-light.js';
 import { linkControllerDemo } from './view-configs/3d-maps/link-controller.js';
+import { linkControllerMinimal } from './view-configs/3d-maps/link-controller-minimal.js';
 
 // Nature Methods figures
 import { citeSeq } from './view-configs/paper-figures/cite-seq.js';
@@ -159,6 +160,7 @@ export const configs = {
   'kiemen-2024': threeMinimal,
   'hakimian-2021': threeMinimalLight,
   'link-controller': linkControllerDemo,
+  'link-controller-minimal': linkControllerMinimal,
 
   // Nature Methods figures
   'figure-cite-seq': citeSeq,
